requirements.txt
Flask==2.3.3
gunicorn
psycopg2-binary
python-dotenv
